% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{calculate_yearly_avg_expenditure}
\alias{calculate_yearly_avg_expenditure}
\title{Calculate Yearly Average Food Expenditure}
\usage{
calculate_yearly_avg_expenditure(data, year)
}
\arguments{
\item{data}{A data frame containing the food expenditure data.}

\item{year}{An integer representing the year for which to calculate the average.}
}
\value{
A numeric value representing the average food expenditure for the specified year.
}
\description{
This function calculates the average food expenditure for a specified year across all entities.
}
