% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{cleaned_food_expenditure}
\alias{cleaned_food_expenditure}
\title{Cleaned Food Expenditure Data}
\format{
A data frame with columns:
\describe{
\item{Entity}{Country name}
\item{Code}{Country code}
\item{Year}{Year of record}
\item{Total food expenditure}{Total food expenditure per person}
}
}
\usage{
cleaned_food_expenditure
}
\description{
This dataset contains the cleaned data on food expenditure per person per year.
}
\keyword{datasets}
