% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{cleaned_high_food_prices}
\alias{cleaned_high_food_prices}
\title{Cleaned High Food Prices Data}
\format{
A data frame with columns:
\describe{
\item{Entity}{Country name}
\item{Code}{Country code}
\item{Year}{Year of record}
\item{High_Food_Price_Abnormal}{Proportion of countries with abnormally high food prices}
\item{High_Food_Price_Moderate}{Proportion of countries with moderately high food prices}
}
}
\usage{
cleaned_high_food_prices
}
\description{
This dataset contains the proportion of countries recording abnormally high food prices.
}
\keyword{datasets}
