% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{identify_high_price_countries}
\alias{identify_high_price_countries}
\title{Identify Countries with High Food Prices}
\usage{
identify_high_price_countries(data, year, threshold = 50)
}
\arguments{
\item{data}{A data frame containing high food prices data.}

\item{year}{An integer representing the year for which to identify countries.}

\item{threshold}{A numeric threshold (default 50\%) for abnormal prices.}
}
\value{
A vector of countries with abnormally high food prices for the specified year.
}
\description{
This function returns a list of countries with abnormally high food prices in a given year.
}
