% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4.R
\name{summarize_high_food_prices}
\alias{summarize_high_food_prices}
\title{Summarize High Food Prices by Year}
\usage{
summarize_high_food_prices(data)
}
\arguments{
\item{data}{A data frame containing high food prices data.}
}
\value{
A data frame summarizing the average, minimum, and maximum high food price percentages per year.
}
\description{
This function summarizes the high food prices proportion data by year.
}
